+++ NetworkServiceLib/helpers.cpp
@@ -42,7 +42,7 @@ ulong stringToLongIP(std::string &src) {
  * and does not starts or ends at 255
  */
 bool isValidIP(ulong &ip) {
-    const unsigned int IPSize = 32;
+    const ulong IPSize = 32;
     if ( (ip&0xFF)==0 || (ip&0xFF)==0xFF)
         return false;
     if (((ip>>(IPSize-2))&0xFF)==0 || ((ip>>(IPSize-2))&0xFF)==0xFF)
